fix: jump delay
in style.css file you have the animation jump, so this animation need time to play full, and when you didn't done it, you can try "jump" again. Well it will make start function jump, and it's tryes to play animation and with delay remove it, but animation need more than 300ms, so it's 1000ms. The player will like it more if the game reacts faster, so 900ms is fune.